Init version 1.1.0
Add export for TypeSFTPGoApiClient (include Client, Schema,...)
Add feature to choose role when create
