+++ CHANGELOG.md
@@ -1,0 +1,6 @@
+## Version 1.1.0
+### Features
+- Now you can choose role *admin* or *user* when create **SFTPGoClient**. Default is *user* for backward compatibility
+
+## Version 1.0.3
+This is the first stable version
