+++ example/index.ts
@@ -2,7 +2,7 @@ import { createWriteStream } from 'fs';
 import {
   SFTPGoApiClient,
   createAxiosClient,
-} from '../src/lib/sftpgo-api-client';
+} from '../dist/sftpgo-api-client/src/index';
 import { resolve } from 'path';
 
 async function testOnlyApiClient() {
@@ -40,7 +40,7 @@ async function testClassApiClientWithAuth() {
   // Not need auth header, the instance already set it
   const response = await client.axiosClient.download_user_file(
     {
-      path: '/test/api/avatar.jpg',
+      path: '/test/api/bg-login-2.png',
     },
     undefined,
     {
@@ -50,14 +50,44 @@ async function testClassApiClientWithAuth() {
 
   // Save result to local
   const destFile = createWriteStream(
-    resolve(process.cwd(), 'example', 'tmp', 'avatar.jpg')
+    resolve(process.cwd(), 'example', 'tmp', 'bg-login-2.png')
   );
   response.data.pipe(destFile);
+  destFile.on('close', () => {
+    console.log('Test successfull!!!');
+  });
+}
+
+async function testClassApiClientWithAuthAdmin() {
+  const client = new SFTPGoApiClient({
+    createApiClientOption: {
+      serverUrl: 'http://localhost:8080/api/v2',
+    },
+    auth: {
+      username: 'admin',
+      password: '12345678',
+      role: 'admin',
+    },
+  });
+
+  // Call this method before each request to ensure access token
+  await client.ensureToken();
+
+  // Not need auth header, the instance already set it
+  const response = await client.axiosClient.get_admin_profile();
+
+  if (response.data != null) {
+    console.log('Test successfull!!!');
+    console.log({
+      data: response.data,
+    });
+  }
 }
 
 async function test() {
   await testOnlyApiClient();
   await testClassApiClientWithAuth();
+  await testClassApiClientWithAuthAdmin();
 }
 
 test();
