+++ src/lib/sftpgo-api-client/client.ts
@@ -1,0 +1,2 @@
+export type { Client as TypeSFTPGoApiClientAxios } from '../client';
+export type * as TypeSFTPGoApiClient from '../client';
