+++ src/lib/sftpgo-api-client/createDocument.ts
@@ -1,0 +1,15 @@
+import OpenAPIClientAxios, { AxiosRequestConfig } from 'openapi-client-axios';
+
+export type CreateDocumentOptions = {
+  definition: string;
+  axiosConfigDefaults?: AxiosRequestConfig;
+};
+
+export async function createDocument(options: CreateDocumentOptions) {
+  const api = new OpenAPIClientAxios({
+    definition: options.definition,
+    axiosConfigDefaults: options.axiosConfigDefaults,
+  });
+  const document = await api.loadDocument();
+  return document;
+}
